% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_anova_table}
\alias{add_anova_table}
\title{Adds a stats::anova(fit1, fit2) table.}
\usage{
add_anova_table(x, ...)
}
\arguments{
\item{x}{An anova comparing fit1 to fit2.}

\item{...}{Args for add_table().}
}
\value{
An Anova flextable.
}
\description{
Adds a stats::anova(fit1, fit2) table.
}
