% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apatfa.R
\name{add_appendix}
\alias{add_appendix}
\title{Adds an appendix}
\usage{
add_appendix(bookmark, fun, wide = FALSE)
}
\arguments{
\item{bookmark}{Bookmark for the appendix}

\item{fun}{Function to add the appendix content to an rdocx object}

\item{wide}{Should the appendix be displayed in landscape?}
}
\description{
Adds an appendix
}
