% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apatfa.R
\name{add_code_file}
\alias{add_code_file}
\title{Adds the text of a code file in a mono 10pt font, single-spaced}
\usage{
add_code_file(x, file_name)
}
\arguments{
\item{x}{the rdocx document}

\item{file_name}{the name of the code file}
}
\value{
the rdocx document
}
\description{
Adds the text of a code file in a mono 10pt font, single-spaced
}
