% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apatfa.R
\name{add_figure}
\alias{add_figure}
\title{Adds a figure}
\usage{
add_figure(fig, bookmark, title, styles, ...)
}
\arguments{
\item{fig}{The figure.}

\item{bookmark}{Bookmark for the figure.}

\item{title}{Title for the figure.}

\item{styles}{A styles list to use.}

\item{...}{Args to pass to begin_figure.}
}
\value{
The figure.
}
\description{
Adds a figure
}
