% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_fit_obf_fig}
\alias{add_fit_obf_fig}
\title{Adds an observed versus fitted plot.}
\usage{
add_fit_obf_fig(fit, bookmark, title)
}
\arguments{
\item{fit}{A fit}

\item{bookmark}{Bookmark for the figure}

\item{title}{Title for the figure}
}
\value{
A figure.
}
\description{
Adds an observed versus fitted plot.
}
