% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_fit_qq_fig}
\alias{add_fit_qq_fig}
\title{Adds a normal Q-Q plot of residuals.}
\usage{
add_fit_qq_fig(fit, bookmark, title, styles)
}
\arguments{
\item{fit}{A fit.}

\item{bookmark}{Bookmark for the figure}

\item{title}{Title for the figure}

\item{styles}{A styles list to use.}
}
\value{
A figure.
}
\description{
Outliers among residuals are labeled.
}
