% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_lm_table}
\alias{add_lm_table}
\title{Adds an lm table.}
\usage{
add_lm_table(x, bookmark, title, styles, notes = NULL, ...)
}
\arguments{
\item{x}{An lm fit.}

\item{bookmark}{Bookmark for the table.}

\item{title}{Title for the table.}

\item{styles}{A styles list to use.}

\item{notes}{Notes about the table.}

\item{...}{Additional args for add_table().}
}
\value{
A flextable.
}
\description{
Adds an lm table.
}
