% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apatfa.R
\name{add_md}
\alias{add_md}
\title{Adds markdown content as paragraphs in the default style.}
\usage{
add_md(x, ...)
}
\arguments{
\item{x}{The rdocx document.}

\item{...}{Markdown content.}
}
\value{
The rdocx document.
}
\description{
Adds markdown content as paragraphs in the default style.
}
