% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apatfa.R
\name{add_md_normal}
\alias{add_md_normal}
\title{Adds markdown content as paragraphs with style = "Normal"}
\usage{
add_md_normal(x, ...)
}
\arguments{
\item{x}{the rdocx document}

\item{...}{markdown content}
}
\value{
the rdocx document
}
\description{
Adds markdown content as paragraphs with style = "Normal"
}
