% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_styling}
\alias{add_styling}
\title{Adds styling for the columns of a data frame.}
\usage{
add_styling(styles, df)
}
\arguments{
\item{styles}{Existing styles.}

\item{df}{The data frame to add.}
}
\value{
Updated styles.
}
\description{
Factor and logical columns will use mono face for values.  All column
names will be in italics.
}
