% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apatfa.R
\name{add_table}
\alias{add_table}
\title{Adds a flextable}
\usage{
add_table(x, bookmark, title, notes = NULL, wide = FALSE)
}
\arguments{
\item{x}{A flextable}

\item{bookmark}{Bookmark for the table}

\item{title}{Title for the table}

\item{notes}{Notes about the table}

\item{wide}{Should the table be displayed in landscape?}
}
\value{
A flextable
}
\description{
Adds a flextable
}
