% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apa_paragraph_md}
\alias{apa_paragraph_md}
\title{Converts markdown to a paragraph in APA style.}
\usage{
apa_paragraph_md(x, styles, ...)
}
\arguments{
\item{x}{A character vector.}

\item{styles}{The styles list to use.}

\item{...}{The args to pass to ftExtra::as_paragraph_md().}
}
\value{
A paragraph.
}
\description{
Calls ftExtra::as_paragraph_md(...) and then applies APA
styles.
}
