% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_flextable.aov}
\alias{as_flextable.aov}
\title{Converts an aov to a flextable.}
\usage{
\method{as_flextable}{aov}(x)
}
\arguments{
\item{x}{An aov.}
}
\value{
A flextable.
}
\description{
Converts an aov to a flextable.
}
