% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_flextable.effectsize_anova}
\alias{as_flextable.effectsize_anova}
\title{Converts an effectsize_anova to a flextable.}
\usage{
\method{as_flextable}{effectsize_anova}(x)
}
\arguments{
\item{x}{An effectsize_anova, such as from effectsize::eta_squared().}
}
\value{
A flextable.
}
\description{
Converts an effectsize_anova to a flextable.
}
