% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_flextable.htest}
\alias{as_flextable.htest}
\title{Converts an htest to a flextable.}
\usage{
\method{as_flextable}{htest}(x)
}
\arguments{
\item{x}{An htest.}
}
\value{
A flextable
}
\description{
Converts an htest to a flextable.
}
