% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_flextable.kruskal_effsize}
\alias{as_flextable.kruskal_effsize}
\title{Converts a kruskal_effsize to a flextable.}
\usage{
\method{as_flextable}{kruskal_effsize}(x)
}
\arguments{
\item{x}{A kruskal_effsize object.}
}
\value{
A flextable.
}
\description{
Converts a kruskal_effsize to a flextable.
}
