% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_flextable.power.htest}
\alias{as_flextable.power.htest}
\title{Converts a power.htest to a flextable.}
\usage{
\method{as_flextable}{power.htest}(x)
}
\arguments{
\item{x}{A power.htest such as from power.anova.test().}
}
\value{
A flextable.
}
\description{
Converts a power.htest to a flextable.
}
