% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_flextable.TukeyHSD}
\alias{as_flextable.TukeyHSD}
\title{Converts a TukeyHSD to a flextable.}
\usage{
\method{as_flextable}{TukeyHSD}(x)
}
\arguments{
\item{x}{A TukeyHSD.}
}
\value{
A flextable.
}
\description{
Converts a TukeyHSD to a flextable.
}
