% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_flextable_dwt}
\alias{as_flextable_dwt}
\title{Converts a Durbin-Watson Test to a flextable.}
\usage{
as_flextable_dwt(x)
}
\arguments{
\item{x}{A dwt test.}
}
\value{
A flextable.
}
\description{
Converts a Durbin-Watson Test to a flextable.
}
