% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_flextable_glm}
\alias{as_flextable_glm}
\title{Converts a glm to a flextable.}
\usage{
as_flextable_glm(x, profile = NULL)
}
\arguments{
\item{x}{A glm object.}

\item{profile}{A profile of the glm object from profile.glm().}
}
\value{
A flextable.
}
\description{
Converts a glm to a flextable.
}
