% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_flextable_htest}
\alias{as_flextable_htest}
\title{Converts an htest to a flextable.}
\usage{
as_flextable_htest(x)
}
\arguments{
\item{x}{An htest.}
}
\value{
A flextable
}
\description{
Converts an htest to a flextable.
}
