% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_flextable_leveneTest}
\alias{as_flextable_leveneTest}
\title{Converts a Levene Test to a flextable.}
\usage{
as_flextable_leveneTest(x)
}
\arguments{
\item{x}{A Levene test.}
}
\value{
A flextable.
}
\description{
Converts a Levene Test to a flextable.
}
