% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_flextable_lm}
\alias{as_flextable_lm}
\title{Converts an lm to a flextable.}
\usage{
as_flextable_lm(x)
}
\arguments{
\item{x}{An lm object.}
}
\value{
A flextable.
}
\description{
Converts an lm to a flextable.
}
