% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_phrase}
\alias{as_phrase}
\title{Converts a vector of strings to a phrase.}
\usage{
as_phrase(g)
}
\arguments{
\item{g}{A vector of strings.}
}
\value{
A phrase.
}
\description{
Converts a vector of strings to a phrase.
}
\examples{
as_phrase(c("A", "B", "C"))
}
