% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_t}
\alias{as_t}
\title{Converts a string to a chunk of text in the default table font.}
\usage{
as_t(x, props = NULL)
}
\arguments{
\item{x}{A string.}

\item{props}{The fp_text properties to use (NULL for default).}
}
\value{
A chunk of text.
}
\description{
Converts a string to a chunk of text in the default table font.
}
