% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apatfa.R
\name{autofit_width}
\alias{autofit_width}
\title{Fits the column widths of a flextable using dim_pretty}
\usage{
autofit_width(x, part = c("body", "header"))
}
\arguments{
\item{x}{A flextable}

\item{part}{The part of the table to use for measuring pretty widths}
}
\value{
A flextable
}
\description{
Fits the column widths of a flextable using dim_pretty
}
