% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apatfa.R
\name{begin_figure}
\alias{begin_figure}
\title{Begins capturing a figure using svg}
\usage{
begin_figure(
  bookmark,
  title,
  wide = FALSE,
  width = NULL,
  height = NULL,
  reserve = 0,
  notes = NULL
)
}
\arguments{
\item{bookmark}{Bookmark for the figure}

\item{title}{Title for the figure}

\item{wide}{Should the figure be displayed in landscape?}

\item{width}{Width for the figure in inches}

\item{height}{Height for the figure in inches}

\item{reserve}{Amount to subtract from the height}

\item{notes}{Notes about the figure (string, fpar, or block_list)}
}
\description{
Begins capturing a figure using svg
}
