% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{blank_axis_x}
\alias{blank_axis_x}
\title{Conditionally blanks the x axis.}
\usage{
blank_axis_x(cond = TRUE)
}
\arguments{
\item{cond}{The condition when the x axis should be blank.}
}
\value{
A theme that either blanks the x axis or does not.
}
\description{
Conditionally blanks the x axis.
}
