% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{blank_axis_y}
\alias{blank_axis_y}
\title{Conditionally blanks the y axis.}
\usage{
blank_axis_y(cond = TRUE)
}
\arguments{
\item{cond}{The condition when the x axis should be blank.}
}
\value{
A theme that either blanks the x axis or does not.
}
\description{
Conditionally blanks the y axis.
}
