% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dstats}
\alias{dstats}
\title{Gets descriptive statistics for variable v.}
\usage{
dstats(v)
}
\arguments{
\item{v}{The values to describe}
}
\value{
A tibble of descriptive stats
}
\description{
Gets descriptive statistics for variable v.
}
