% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ggimg}
\alias{ggimg}
\title{Converts a static image to a ggplot with fixed coordinates.}
\usage{
ggimg(img)
}
\arguments{
\item{img}{An image, such as from png::readPNG}
}
\value{
A ggplot of the image
}
\description{
Converts a static image to a ggplot with fixed coordinates.
}
