% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grouped_bookmark}
\alias{grouped_bookmark}
\title{Gets a bookmark for a grouping combination.}
\usage{
grouped_bookmark(bookmark, g)
}
\arguments{
\item{bookmark}{The base bookmark.}

\item{g}{A vector of grouping variable names.}
}
\value{
The base bookmark suffixed with grouping information.
}
\description{
Gets a bookmark for a grouping combination.
}
