% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grouped_combn}
\alias{grouped_combn}
\title{Converts a vector of grouping variables to a list of combinations
of the grouping variables taken m at a time for each m in range r.}
\usage{
grouped_combn(gv, r = seq.int(0, length(gv)))
}
\arguments{
\item{gv}{A vector of group variable names.}

\item{r}{The combinations range, defaults to all combinations.}
}
\value{
A list of combinations of the grouping variables.
}
\description{
Converts a vector of grouping variables to a list of combinations
of the grouping variables taken m at a time for each m in range r.
}
