% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grouped_title}
\alias{grouped_title}
\title{Gets a title for a grouping combination.}
\usage{
grouped_title(title, g)
}
\arguments{
\item{title}{The base title.}

\item{g}{A vector of grouping variable names.}
}
\value{
The base title suffixed with grouping information.
}
\description{
Gets a title for a grouping combination.
}
