% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hrotate}
\alias{hrotate}
\title{Rotates the header of a flextable.}
\usage{
hrotate(x)
}
\arguments{
\item{x}{A flextable}
}
\value{
The flextable with rotated headers
}
\description{
Rotates the header of a flextable.
}
