% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{integers_in_extended_range}
\alias{integers_in_extended_range}
\title{Returns the integer values in the range of x, expanded.}
\usage{
integers_in_extended_range(x)
}
\arguments{
\item{x}{The axis values.}
}
\value{
A sequence of values in the range.
}
\description{
Returns the integer values in the range of x, expanded.
}
