% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{integers_in_range}
\alias{integers_in_range}
\title{Returns the integer values in the range of x.}
\usage{
integers_in_range(x, extend = FALSE)
}
\arguments{
\item{x}{The axis values.}

\item{extend}{If TRUE, lean out on the borders.}
}
\value{
A sequence of values in the range.
}
\description{
Returns the integer values in the range of x.
}
