% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_normal}
\alias{is_normal}
\title{Runs a Shapiro-Wilk test of normality.}
\usage{
is_normal(v, alpha = 0.05)
}
\arguments{
\item{v}{The values to test.}

\item{alpha}{The alpha level.}
}
\description{
Returns "Yes" if v is normal, "No" if v is not normal,
or "NA" if the test cannot be run.
}
