% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_outlier}
\alias{is_outlier}
\title{Uses the geom_boxplot() algorithm to identify the outlier points.}
\usage{
is_outlier(x, ...)
}
\arguments{
\item{x}{A vector of numeric data.}

\item{...}{Additional parameters to pass to geom_boxplot().}
}
\value{
A logical vector the same length as x (TRUE if outlier).
}
\description{
Uses the geom_boxplot() algorithm to identify the outlier points.
}
