% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md.R
\name{md_notes}
\alias{md_notes}
\title{Creates notes from markdown.}
\usage{
md_notes(...)
}
\arguments{
\item{...}{markdown content}
}
\value{
an officer::block_list of paragraphs
}
\description{
Creates notes from markdown.
}
