% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md.R
\name{md_title}
\alias{md_title}
\title{Creates a title from markdown.}
\usage{
md_title(...)
}
\arguments{
\item{...}{markdown content}
}
\value{
an officer::block_list of paragraphs
}
\description{
Creates a title from markdown.
}
