% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{n_is}
\alias{n_is}
\title{Appends a suffix regarding the size of a data frame.}
\usage{
n_is(title = NULL, df, n = "n")
}
\arguments{
\item{df}{The data frame.}

\item{n}{The name to use to label the size, usually "n" or "N".}
}
\description{
Appends a suffix regarding the size of a data frame.
}
