% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normal_note}
\alias{normal_note}
\title{Appends a note about a Shapiro-Wilk test.}
\usage{
normal_note(notes = NULL, alpha = 0.05, what = "Coloring")
}
\arguments{
\item{notes}{Previous notes.}

\item{alpha}{The alpha level used for testing.}

\item{what}{The aesthetic used to indicate the test result.}
}
\description{
Appends a note about a Shapiro-Wilk test.
}
