% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{note_cor_test}
\alias{note_cor_test}
\title{Performs a correlation test and appends a note about it.}
\usage{
note_cor_test(notes, df, x, y, alpha = 0.05, ...)
}
\arguments{
\item{notes}{The notes.}

\item{df}{A data frame.}

\item{x}{The column name for the x variable.}

\item{y}{The column name for the y variable.}

\item{alpha}{The alpha level for the correlation test.}

\item{...}{Additional args for stats::cor.test().}
}
\value{
The notes, with a correlation test note appended.
}
\description{
Performs a correlation test and appends a note about it.
}
