% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{note_facets}
\alias{note_facets}
\title{Appends a note about faceting.}
\usage{
note_facets(notes = NULL, facet_vars)
}
\arguments{
\item{notes}{The notes.}

\item{facet_vars}{A vector of facet variable names.}
}
\value{
The notes, with a note about faceting appended.
}
\description{
Appends a note about faceting.
}
