% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{note_fit_model}
\alias{note_fit_model}
\title{Appends a note about that.}
\usage{
note_fit_model(notes = NULL, fit)
}
\arguments{
\item{notes}{The notes.}

\item{fit}{The model of this fit will be noted.}
}
\value{
The updated notes.
}
\description{
Appends a note about that.
}
