% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{note_intro}
\alias{note_intro}
\title{Prefixes notes with an intro: Note.}
\usage{
note_intro(notes)
}
\arguments{
\item{notes}{The notes.}
}
\value{
The notes, prefixed if the intro was missing.
}
\description{
Prefixes notes with an intro: Note.
}
