% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{note_normal}
\alias{note_normal}
\title{Appends a note about a Shapiro-Wilk test of normality.}
\usage{
note_normal(notes = NULL, what = "Coloring", alpha = 0.05)
}
\arguments{
\item{notes}{Previous notes.}

\item{what}{The aesthetic used to indicate the test result.}

\item{alpha}{The alpha level used for testing.}
}
\value{
The previous notes with a normality note appended.
}
\description{
Appends a note about a Shapiro-Wilk test of normality.
}
