% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{note_p_levels}
\alias{note_p_levels}
\title{Returns a paragraph of significance levels for p.}
\usage{
note_p_levels()
}
\value{
A paragraph.
}
\description{
Returns a paragraph of significance levels for p.
}
