% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apatfa.R
\name{note_paras}
\alias{note_paras}
\title{Converts a list of strings to a list of styled paragraphs.}
\usage{
note_paras(notes, styles, as_title = FALSE)
}
\arguments{
\item{styles}{The styles to apply.}

\item{as_title}{If TRUE, apply inverse italics for titles.}

\item{paras}{A list of strings.}
}
\value{
A list of styled paragraphs.
}
\description{
Converts a list of strings to a list of styled paragraphs.
}
