% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{note_statistic_htest}
\alias{note_statistic_htest}
\title{Appends a note about an statistic from a hypothesis test.}
\usage{
note_statistic_htest(notes, h)
}
\arguments{
\item{notes}{The notes.}

\item{h}{A hypothesis test.}
}
\value{
The notes, with a note about the statistic appended.
}
\description{
Appends a note about an statistic from a hypothesis test.
}
