% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apatfa.R
\name{note_table}
\alias{note_table}
\title{Converts a list of paragraphs to a flextable with one column.}
\usage{
note_table(paras)
}
\arguments{
\item{paras}{A list of paragraphs.}
}
\value{
A flextable.
}
\description{
Converts a list of paragraphs to a flextable with one column.
}
