% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{note_that}
\alias{note_that}
\title{Appends a note about that.}
\usage{
note_that(notes = NULL, ..., with_dot = TRUE)
}
\arguments{
\item{notes}{The notes.}

\item{...}{That to note.}

\item{with_dot}{If TRUE, the note will end in a period.}
}
\value{
The notes, with a note about that appended.
}
\description{
Appends a note about that.
}
