% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{out_of_date}
\alias{out_of_date}
\title{Performs a dependency check.}
\usage{
out_of_date(target, dependency)
}
\arguments{
\item{target}{The target file.}

\item{dependency}{The dependency file.}
}
\value{
TRUE if the target is out of date.
}
\description{
Performs a dependency check.
}
