% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{poisson_lambda_ci}
\alias{poisson_lambda_ci}
\title{Returns the estimated Poisson lambda parameter with CI.}
\usage{
poisson_lambda_ci(x, ...)
}
\arguments{
\item{x}{A numeric vector of data.}

\item{...}{Other args for poisson.exact.}
}
\value{
A tibble with ymin, y, and ymax columns.
}
\description{
Returns the estimated Poisson lambda parameter with CI.
}
