% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pval}
\alias{pval}
\title{Converts p values to formatted text.}
\usage{
pval(p, with_p = TRUE, with_eq = with_p)
}
\arguments{
\item{p}{The p values.}

\item{with_p}{Logical.  Prefix with the p character?}

\item{with_eq}{Logical.  Use equal sign?}
}
\value{
Formatted paragraphs.
}
\description{
Converts p values to formatted text.
}
