% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apatfa.R
\name{rotate_header}
\alias{rotate_header}
\title{Rotates and aligns the header of a flextable}
\usage{
rotate_header(x, rotation = "tbrl", align = "right")
}
\arguments{
\item{x}{A flextable}

\item{rotation}{Header rotation (tbrl or btlr)}

\item{align}{Header alignment (left, center, or right)}
}
\value{
A flextable
}
\description{
Rotates and aligns the header of a flextable
}
