% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_fill_yes_no_na}
\alias{scale_fill_yes_no_na}
\title{Gets a scale_fill_manual with Yes, No, NA aesthetic keys.}
\usage{
scale_fill_yes_no_na(styles = get_styles())
}
\arguments{
\item{styles}{Can use custom colors from a styles list.}
}
\value{
A scale_fill_manual object
}
\description{
Gets a scale_fill_manual with Yes, No, NA aesthetic keys.
}
