% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apatfa.R
\name{set_apa_defaults}
\alias{set_apa_defaults}
\title{Sets flextable defaults for APA style tables}
\usage{
set_apa_defaults(...)
}
\arguments{
\item{...}{Args to pass to set_flextable_defaults()}
}
\value{
A list containing the previous default values
}
\description{
Sets flextable defaults for APA style tables
}
