% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{signif_format}
\alias{signif_format}
\title{Returns significance codes using asterisk and dagger characters.}
\usage{
signif_format(x)
}
\arguments{
\item{x}{The p values.}
}
\value{
Significance codes.
}
\description{
Returns significance codes using asterisk and dagger characters.
}
