% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{StamenAttribution}
\alias{StamenAttribution}
\title{Creates an attribution label for maps provided by Stamen Design.}
\usage{
StamenAttribution(x, y, size = 3)
}
\arguments{
\item{x}{The x coordinate for the right edge of the label.}

\item{y}{The y coordinate for the bottom edge of the label.}

\item{size}{The label size.}
}
\value{
A geom_label.
}
\description{
Creates an attribution label for maps provided by Stamen Design.
}
