% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{styler}
\alias{styler}
\title{Applies styles to a flextable.}
\usage{
styler(x, styles, i = NULL, fit = TRUE)
}
\arguments{
\item{x}{A flextable.}

\item{styles}{A styles list to use.}

\item{i}{A column selector for the flextable.}

\item{fit}{Fit to the width if TRUE.}
}
\value{
The styled flextable.
}
\description{
Applies styles to a flextable.
}
