% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{styler_with_spanner}
\alias{styler_with_spanner}
\title{Applies styles to a flextable with a table spanner.}
\usage{
styler_with_spanner(x, styles, spanner)
}
\arguments{
\item{x}{A flextable.}

\item{styles}{A styles list to use.}

\item{spanner}{The spanner variable name.}
}
\value{
The styled flextable.
}
\description{
Applies styles to a flextable with a table spanner.
}
