% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tcase}
\alias{tcase}
\title{Converts text to APA title case.}
\usage{
tcase(t)
}
\arguments{
\item{t}{The text to convert.}
}
\value{
The text in title case
}
\description{
Converts text to APA title case.
}
