% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apatfa.R
\name{theme_apa}
\alias{theme_apa}
\title{Applies the APA theme to a flextable}
\usage{
theme_apa(x)
}
\arguments{
\item{x}{Input flextable}
}
\value{
Output flextable in APA style
}
\description{
The APA theme is similar to the booktabs theme, except:
\itemize{
\item all headers are center aligned
\item the left body column is left aligned
\item all other body columns are center aligned
\item numbers are padded to align on the decimal point
}
}
