% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{title_case}
\alias{title_case}
\title{Converts a title to APA title case.}
\usage{
title_case(title)
}
\arguments{
\item{title}{The title to convert.}
}
\value{
The title in title case.
}
\description{
Converts a title to APA title case.
}
