% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{title_n}
\alias{title_n}
\title{Suffixes a title with information about the number of observations.}
\usage{
title_n(title = NULL, df, n = "n")
}
\arguments{
\item{title}{The title.}

\item{df}{The data frame.}

\item{n}{The name to use to label the size, usually "n" or "N".}
}
\description{
Suffixes a title with information about the number of observations.
}
