% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zscore}
\alias{zscore}
\title{Computes standardized scores.}
\usage{
zscore(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector or an \R object but not a
    \code{\link{factor}} coercible to numeric by \code{as.double(x)}.}

\item{na.rm}{logical.  Should missing values be removed?}
}
\value{
The standardized scores for the values.
}
\description{
Subtracts the mean and then divides by the standard deviation.
}
