% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_anova_table}
\alias{add_anova_table}
\title{Adds a car::Anova() table.}
\usage{
add_anova_table(x, ...)
}
\arguments{
\item{x}{An Anova.}

\item{...}{Args for add_table().}
}
\value{
An Anova flextable.
}
\description{
Adds a car::Anova() table.
}
