% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apatfa.R
\name{add_code_file}
\alias{add_code_file}
\title{Adds the text of a code file in a mono 10pt font, single-spaced.}
\usage{
add_code_file(x, file_name, head = FALSE)
}
\arguments{
\item{x}{The rdocx document.}

\item{file_name}{The name of the code file.}

\item{head}{If TRUE, only add the first few lines of the file.}
}
\value{
The rdocx document.
}
\description{
Adds the text of a code file in a mono 10pt font, single-spaced.
}
