% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apatfa.R
\name{add_figure}
\alias{add_figure}
\title{Adds a figure}
\usage{
add_figure(
  fig,
  bookmark,
  title,
  styles,
  notes = NULL,
  wide = FALSE,
  width = NULL,
  height = NULL,
  reserve = 0
)
}
\arguments{
\item{fig}{The figure}

\item{bookmark}{Bookmark for the figure}

\item{title}{Title for the figure}

\item{styles}{A styles list to use.}

\item{notes}{Notes about the figure}

\item{wide}{Should the figure be displayed in landscape?}

\item{width}{Width for the figure in inches}

\item{height}{Height for the figure in inches}

\item{reserve}{Amount to subtract from the height}
}
\value{
The figure
}
\description{
Adds a figure
}
