% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_fit_rp_fig}
\alias{add_fit_rp_fig}
\title{Adds a plot of residual by predicted value.}
\usage{
add_fit_rp_fig(fit, bookmark, title, type = "pearson")
}
\arguments{
\item{fit}{A fit}

\item{bookmark}{Bookmark for the figure}

\item{title}{Title for the figure}

\item{type}{The type of residual to plot.}
}
\value{
A figure.
}
\description{
Adds a plot of residual by predicted value.
}
