% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_glm_table}
\alias{add_glm_table}
\title{Adds a glm table.}
\usage{
add_glm_table(x, styles, ...)
}
\arguments{
\item{x}{A glm fit.}

\item{styles}{A styles list to use.}

\item{...}{Args for add_table().}
}
\value{
A flextable.
}
\description{
Adds a glm table.
}
