% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_lm_table}
\alias{add_lm_table}
\title{Adds an lm table.}
\usage{
add_lm_table(x, ...)
}
\arguments{
\item{x}{An lm fit.}

\item{...}{Args for add_table().}
}
\value{
A flextable.
}
\description{
Adds an lm table.
}
