% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apatfa-package.R
\docType{package}
\name{apatfa-package}
\alias{apatfa-package}
\alias{_PACKAGE}
\title{apatfa: Generates APA Tables, Figures, and Appendix Sections}
\description{
Generates an output Word file (docx) containing bookmarked content (tables, figures, and appendices) numbered and organized in APA style sections. In each output section, the content is ordered by the first reference to the content within an input Word file. The output file is designed to be included into the input file using an "IncludeText" field in Word so that bookmarks in the output file can be referenced from the input file. This package primarily uses the officer and flextable packages, but the difference here is the obsession with conforming to the APA style guide.
}
\details{
For more details see the README file at
\url{https://github.com/toddagood/apatfa/}.
}
\seealso{
\url{https://github.com/toddagood/apatfa/}

\url{https://github.com/davidgohel/flextable/}

\url{https://github.com/davidgohel/officer/}
}
\author{
\strong{Maintainer}: Todd Goodrich \email{toddagood@gmail.com}

}
\keyword{internal}
