% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apatfa.R
\name{begin_figure}
\alias{begin_figure}
\title{Begins capturing a figure using svg}
\usage{
begin_figure(
  bookmark,
  title,
  styles,
  notes = NULL,
  wide = FALSE,
  width = NULL,
  height = NULL,
  reserve = 0
)
}
\arguments{
\item{bookmark}{Bookmark for the figure.}

\item{title}{Title for the figure.}

\item{styles}{A styles list to use.}

\item{notes}{Notes about the figure.}

\item{wide}{Should the figure be displayed in landscape?}

\item{width}{Width for the figure in inches.}

\item{height}{Height for the figure in inches.}

\item{reserve}{Amount to subtract from the height.}
}
\description{
The title and notes args accept flextable objects.
If a string is provided instead it will be converted
to a flextable with one row and italics based on the styles.
If a list of strings is provided instead it will be converted
to a flextable with a row for each list item.
}
