% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dstats_row}
\alias{dstats_row}
\title{Gets a tibble containing the name of the variable (in the first column)
and descriptive statistics for the named variable in the remaining
columns}
\usage{
dstats_row(name, df)
}
\arguments{
\item{name}{The variable name}

\item{df}{The data frame}
}
\value{
A tibble of descriptive stats
}
\description{
Gets a tibble containing the name of the variable (in the first column)
and descriptive statistics for the named variable in the remaining
columns
}
