% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{note_p_value}
\alias{note_p_value}
\title{Appends a note about a p value.}
\usage{
note_p_value(notes = NULL, p, with_p = TRUE, with_eq = with_p)
}
\arguments{
\item{notes}{The notes.}

\item{p}{The p value.}

\item{with_p}{Logical.  Prefix with the p character?}

\item{with_eq}{Logical.  Use equal sign?}
}
\value{
The notes, with a note about a p value appended.
}
\description{
Appends a note about a p value.
}
