% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{note_that}
\alias{note_that}
\title{Appends a note about that.}
\usage{
note_that(notes, that)
}
\arguments{
\item{notes}{The notes.}

\item{that}{That to note.}
}
\value{
The notes, with a note about that appended.
}
\description{
Appends a note about that.
}
