% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pval_pars}
\alias{pval_pars}
\title{Converts p values to formatted paragraphs.}
\usage{
pval_pars(pvals)
}
\arguments{
\item{pvals}{The p values}
}
\value{
Formatted paragraphs
}
\description{
Converts p values to formatted paragraphs.
}
