% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apatfa.R
\name{set_apa_defaults}
\alias{set_apa_defaults}
\title{Sets APA style defaults for flextable and ggplot2}
\usage{
set_apa_defaults(
  digits = 2,
  na_str = "NA",
  nan_str = "NaN",
  fmt_datetime = "\%Y-\%m-\%d \%H:\%M:\%S \%Z",
  fig_theme = ggplot2::theme(),
  styles = get_styles(),
  ...
)
}
\arguments{
\item{digits}{significant digits for tables}

\item{na_str}{string to be used for NA values in tables}

\item{nan_str}{string to be used for NaN values in tables}

\item{fmt_datetime}{format string for datatime values}

\item{fig_theme}{additional ggplot2::theme settings for figures}

\item{styles}{Styles to use.}

\item{...}{args to pass to set_flextable_defaults()}
}
\description{
Sets APA style defaults for flextable and ggplot2
}
